Adds axios for header management and forces middleware to return users
Now the authenticate and isLoggedIn functions return users and the
router object has a way of formatting them. Additionally, data is now
read from local storage properly and fed into the headers if it contains
a valid token. User data is never stored and is fetched when validating
a token or logging in to ensure that it is always up to date.